refactor: move restore impl to DwcMaskedField
